Refactor classifier_nn tests into separate directory
